requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.14
django-celery-beat==2.3.0
django-crispy-forms==1.14.0
django-timezone-field==5.0
djangorestframework==3.13.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
prompt-toolkit==3.0.30
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
zope.interface==5.4.0
